Rename MuxerFilter into Muxer. Rename Muxer into Engine and put it in Muxer. namespace. Update rest of the code accordingly
